Refactoring OperationalStrategy
