+++ org.emoflon.ibex.common/META-INF/MANIFEST.MF
@@ -8,11 +8,11 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: IBeXLanguage,
  IBeXLanguage.impl,
  IBeXLanguage.util,
+ org.emoflon.ibex.common.operational,
  org.emoflon.ibex.common.utils
 Bundle-Vendor: %providerName
 Bundle-Localization: plugin
 Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.emf.ecore;visibility:=reexport,
- org.eclipse.emf.ecore.xmi,
- org.junit
+ org.eclipse.emf.ecore.xmi
 Bundle-ClassPath: .
