+++ org.emoflon.ibex.common/src/org/emoflon/ibex/common/operational/IMatchObserver.java
@@ -1,0 +1,9 @@
+package org.emoflon.ibex.common.operational;
+
+public interface IMatchObserver {
+	public void addMatch(IMatch match);
+
+	public void removeMatch(IMatch match);
+
+	public boolean isPatternRelevantForCompiler(String patternName);
+}
