+++ org.emoflon.ibex.tgg.core.runtime/META-INF/MANIFEST.MF
@@ -14,7 +14,8 @@ Require-Bundle: org.eclipse.emf.ecore,
  org.eclipse.core.runtime,
  org.eclipse.core.resources,
  org.emoflon.ibex.tgg.editor,
- org.emoflon.ibex.tgg.ide
+ org.emoflon.ibex.tgg.ide,
+ org.emoflon.ibex.common
 Bundle-ManifestVersion: 2
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
