+++ org.emoflon.ibex.tgg.core.runtime/src/org/emoflon/ibex/tgg/operational/matches/IMatch.java
@@ -4,21 +4,7 @@ import java.util.Collection;
 
 import org.emoflon.ibex.tgg.operational.edge.RuntimeEdge;
 
-public interface IMatch {
-
-	Object get(String name);
-	
-	default void put(String name, Object o) {
-		throw new UnsupportedOperationException("This match does not support adding new mappings!");
-	}
-
-	Collection<String> parameterNames();
-
-	String patternName();
-
-	default boolean isInMatch(String name) {
-		return parameterNames().contains(name);
-	}
+public interface IMatch extends org.emoflon.ibex.common.operational.IMatch {
 
 	IMatch copy();
