+++ org.emoflon.ibex.tgg.core.runtime/src/org/emoflon/ibex/tgg/operational/strategies/OperationalStrategy.java
@@ -19,6 +19,7 @@ import org.eclipse.emf.ecore.resource.ContentHandler;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.ResourceSet;
 import org.eclipse.emf.ecore.util.EcoreUtil;
+import org.emoflon.ibex.common.operational.IMatchObserver;
 import org.emoflon.ibex.tgg.compiler.patterns.PatternSuffixes;
 import org.emoflon.ibex.tgg.compiler.patterns.sync.ConsistencyPattern;
 import org.emoflon.ibex.tgg.operational.IBlackInterpreter;
@@ -54,7 +55,7 @@ import runtime.TGGRuleApplication;
 import runtime.TempContainer;
 import runtime.impl.RuntimePackageImpl;
 
-public abstract class OperationalStrategy {
+public abstract class OperationalStrategy implements IMatchObserver {
 
 	protected final static Logger logger = Logger.getLogger(OperationalStrategy.class);
 	protected final URI base;
@@ -109,6 +110,29 @@ public abstract class OperationalStrategy {
 		greenInterpreter = new IbexGreenInterpreter(this);
 		redInterpreter = new IbexRedInterpreter(this);
 	}
+	
+	@Override
+	public void addMatch(org.emoflon.ibex.common.operational.IMatch match) {
+		this.addOperationalRuleMatch(PatternSuffixes.removeSuffix(match.patternName()), (IMatch) match);
+	}
+	
+	@Override
+	public void removeMatch(org.emoflon.ibex.common.operational.IMatch match) {
+		if (match.patternName().endsWith(PatternSuffixes.CONSISTENCY)) {
+			this.addBrokenMatch((IMatch) match);
+		}
+		this.removeOperationalRuleMatch((IMatch) match);
+	}
+	
+	/**
+	 * Decide if matches of this pattern should be watched and notified by the
+	 * pattern matcher
+	 * 
+	 * @param patternName
+	 * @return
+	 */
+	@Override
+	abstract public boolean isPatternRelevantForCompiler(String patternName);
 
 	protected void setModelGen() {
 		options.setModelGen(false);
@@ -137,15 +161,6 @@ public abstract class OperationalStrategy {
 	}
 
 	protected abstract void registerUserMetamodels() throws IOException;
-
-	/**
-	 * Decide if matches of this pattern should be watched and notified by the
-	 * pattern matcher
-	 * 
-	 * @param patternName
-	 * @return
-	 */
-	abstract public boolean isPatternRelevantForCompiler(String patternName);
 
 	public boolean isPatternRelevantForInterpreter(String patternName) {
 		return isPatternRelevantForCompiler(patternName);
@@ -191,7 +206,7 @@ public abstract class OperationalStrategy {
 	private boolean nodeIsInResource(IMatch match, String name, Resource r) {
 		return ((EObject) match.get(name)).eResource().equals(r);
 	}
-
+	
 	public void removeOperationalRuleMatch(IMatch match) {
 		operationalMatchContainer.removeMatch(match);
 	}
